refactored location of DataFrame tests
