+++ src/testframework/dataquality/dataframe/tests/valid_category.py
@@ -2,7 +2,7 @@ from pyspark.sql import Column, DataFrame
 from pyspark.sql import functions as F
 from pyspark.sql.types import StringType
 
-from testframework.base import Test
+from testframework.dataquality._base import Test
 from testframework.utils.decorators import account_for_nullable, allowed_col_types
