+++ src/testframework/dataquality/dataframe/tests/valid_numeric_range.py
@@ -12,7 +12,7 @@ from pyspark.sql.types import (
     ShortType,
 )
 
-from testframework.base import Test
+from testframework.dataquality._base import Test
 from testframework.utils.decorators import account_for_nullable, allowed_col_types
