+++ tests/dataquality/tests/test_IntegerString.py
@@ -1,6 +1,6 @@
 import pytest
 from pyspark.sql import types as T
-from testframework.dataquality.tests import IntegerString
+from testframework.dataquality.dataframe.tests import IntegerString
 
 test_data = [
     ("01", False, "pk1"),  # Leading zero
